Automate updating help messages for commands
This commit is the first phase of a plan to automate updating our
documentation. It adds a script that can be used to update the help
messages for commands. Instead of copy pasting `--help` output into the
documentation, we can now use this script to generate these messages for
us. All that is left is to include relevant parts of these generated
output in the documentation.